@@ -23,17 +23,17 @@
   "private": true,
   "dependencies": {
     "@algolia/client-search": "^5.23.2",
-    "@angular/animations": "19.2.5",
-    "@angular/cdk": "19.2.8",
-    "@angular/common": "19.2.5",
-    "@angular/compiler": "19.2.5",
-    "@angular/core": "19.2.5",
-    "@angular/forms": "19.2.5",
-    "@angular/platform-browser": "19.2.5",
-    "@angular/platform-browser-dynamic": "19.2.5",
-    "@angular/platform-server": "19.2.5",
-    "@angular/router": "19.2.5",
-    "@angular/ssr": "19.2.6",
+    "@angular/animations": "20.3.10",
+    "@angular/cdk": "20.2.12",
+    "@angular/common": "20.3.10",
+    "@angular/compiler": "20.3.10",
+    "@angular/core": "20.3.10",
+    "@angular/forms": "20.3.10",
+    "@angular/platform-browser": "20.3.10",
+    "@angular/platform-browser-dynamic": "20.3.10",
+    "@angular/platform-server": "20.3.10",
+    "@angular/router": "20.3.10",
+    "@angular/ssr": "20.3.9",
     "@builder.io/partytown": "^0.10.2",
     "@jsverse/transloco": "^7.4.2",
     "@libsql/client": "^0.14.0",
@@ -119,28 +118,29 @@
     "eslint": "8.57.0",
     "eslint-config-prettier": "10.1.1",
     "eslint-plugin-playwright": "^0.15.3",
+    "eslint-plugin-storybook": "9.1.9",
     "husky": "^8.0.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-environment-node": "^29.7.0",
-    "jest-preset-angular": "14.4.2",
+    "jest-preset-angular": "14.6.2",
     "jsonc-eslint-parser": "^2.1.0",
     "lint-staged": "15.0.2",
-    "ng-packagr": "19.2.1",
-    "nx": "20.7.1",
+    "ng-packagr": "20.3.1",
+    "nx": "22.0.3",
     "postcss": "^8.4.32",
     "postcss-import": "~14.1.0",
     "postcss-preset-env": "~7.5.0",
     "postcss-url": "~10.1.3",
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.14",
-    "storybook": "^8.1.0-alpha.7",
+    "storybook": "9.1.9",
     "stylelint": "^16.3.1",
     "svgo": "^3.3.2",
     "tailwindcss": "^3.3.6",
-    "ts-jest": "^29.1.0",
+    "ts-jest": "29.4.5",
     "ts-node": "10.9.1",
-    "typescript": "5.7.3",
+    "typescript": "5.9.3",
     "wrangler": "^3.78.6"
   }
 }
